transfer table type and multi-operations trx handler

+++ app/plugins/dbtables/transferinfo.sql
@@ -4,7 +4,7 @@ create table transferinfo
 	create_time int unsigned not null,
 	sender varchar(64) not null,
 	receiver varchar(64) not null,
-	amount int unsigned default 0,
+	amount bigint default 0,
 	memo TEXT ,
 	INDEX transfer_create_time (create_time),
 	INDEX transfer_sender (sender),
